#557 Removed 'code creation' code from visitor into the C and Fortran backends.
