+++ src/psyclone/psyir/backend/c.py
@@ -403,38 +403,51 @@ class CWriter(PSyIRVisitor):
         '''
         raise VisitorError("CodeBlocks can not be translated to C.")
 
-    @property
-    def do_loop_format(self):
-        '''Returns the format for a do/for loop. The format variables will
-        be replaced as follows:
-        0: indentation string
-        1: loop variable
-        2: first loop iteration
-        3: last loop iteration
-        4: step size
-        5: body of the loop
+    def loop_node(self, node):
+        '''This method is called when a Loop instance is found in the
+        PSyIR tree.
 
-        :return: the format of a loop.
+        :param node: a Loop PSyIR node.
+        :type node: :py:class:`psyclone.psyGen.Loop`
+
+        :returns: the loop node converted into a (language specific) string.
         :rtype: str
+
         '''
-        return "{0}for({1}={2}; {1}<={3}; {1}+={4})\n"\
-            "{0}{{\n"\
-            "{5}"\
-            "{0}}}\n"
+        start = self._visit(node.start_expr)
+        stop = self._visit(node.stop_expr)
+        step = self._visit(node.step_expr)
+        variable_name = node.variable_name
 
-    @property
-    def directive_start(self):
-        ''':return: "#pragma ...\n{" - the opening of a directive in C. The
-        string {0} in the result will be replaced with the actual directive.
+        self._depth += 1
+        body = ""
+        for child in node.loop_body:
+            body += self._visit(child)
+        self._depth -= 1
+
+        return "{0}for({1}={2}; {1}<={3}; {1}+={4})\n"\
+               "{0}{{\n"\
+               "{5}"\
+               "{0}}}\n".format(self._nindent, variable_name,
+                                start, stop, step, body)
+
+    def ompdirective_node(self, node):
+        '''This method is called when an OMPDirective instance is found in
+        the PSyIR tree. It returns the opening and closing directives, and
+        the statements in between as a string (depending on the language).
+
+        :param node: a Directive PSyIR node.
+        :type node: :py:class:`psyclone.psyGen.Directive`
+
+        :returns: the Fortran code as a string.
         :rtype: str
+
         '''
         # Note that {{ is replaced with a single { in the format call
-        return "#pragma {0}\n{{\n"
-
-    @property
-    def directive_end(self):
-        ''':return: "}" - the closing of a directive in C.
-        :rtype: str
-        '''
-        # Note that }} is replaced with a single } in the format call
-        return "}}\n"
+        result_list = ["#pragma {0}\n{{\n".format(node.begin_string())]
+        self._depth += 1
+        for child in node.children:
+            result_list.append(self._visit(child))
+        self._depth -= 1
+        result_list.append("}}\n")
+        return "".join(result_list)
